v3.1/earth-online/backend/requirements.txt
fastapi==0.136.1
uvicorn[standard]==0.46.0
sqlalchemy==2.0.49
pydantic==2.13.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
psycopg2-binary==2.9.12
alembic==1.15.2
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
redis==5.2.0

# Testing dependencies
pytest==8.3.4
pytest-cov==6.0.0
pytest-asyncio==0.24.0
httpx==0.28.1
